// web/src/gen/classicfarm/v1/ws/plot/plot_state_pb.ts
// @generated by protoc-gen-es v2.14.0 with parameter "target=ts"
// @generated from file classicfarm/v1/ws/plot/plot_state.proto (package classicfarm.ws.v1.plot, syntax proto3)
/* eslint-disable */

import type { GenEnum, GenFile } from "@bufbuild/protobuf/codegenv2";
import { enumDesc, fileDesc } from "@bufbuild/protobuf/codegenv2";

/**
 * Describes the file classicfarm/v1/ws/plot/plot_state.proto.
 */
export const file_classicfarm_v1_ws_plot_plot_state: GenFile = /*@__PURE__*/
  fileDesc("CidjbGFzc2ljZmFybS92MS93cy9wbG90L3Bsb3Rfc3RhdGUucHJvdG8SFmNsYXNzaWNmYXJtLndzLnYxLnBsb3QqUgoJUGxvdFN0YXRlEg8KC1VOU1BFQ0lGSUVEEAASCQoFRU1QVFkQARILCgdHUk9XSU5HEAISCgoGTUFUVVJFEAMSEAoMTkVFRF9DTEVBTlVQEARCVVpTZ2l0aHViLmNvbS9Xcmlvc2xleS9zdXBlcm5vdmEtY2xhc3NpYy1mYXJtL3NlcnZlci9nZW4vY2xhc3NpY2Zhcm0vdjEvd3MvcGxvdDtwbG90djFiBnByb3RvMw");

/**
 * @generated from enum classicfarm.ws.v1.plot.PlotState
 */
export enum PlotState {
  /**
   * @generated from enum value: UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: EMPTY = 1;
   */
  EMPTY = 1,

  /**
   * @generated from enum value: GROWING = 2;
   */
  GROWING = 2,

  /**
   * @generated from enum value: MATURE = 3;
   */
  MATURE = 3,

  /**
   * @generated from enum value: NEED_CLEANUP = 4;
   */
  NEED_CLEANUP = 4,
}

/**
 * Describes the enum classicfarm.ws.v1.plot.PlotState.
 */
export const PlotStateSchema: GenEnum<PlotState> = /*@__PURE__*/
  enumDesc(file_classicfarm_v1_ws_plot_plot_state, 0);